pyproject.toml
[project]
name = "langraphcourse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.14",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.4",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
]
